requirements.txt
clarifai
streamlit
pillow
